requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2024.12.14
cffi==2.0.0
charset-normalizer==3.4.1
click==8.2.1
comm==0.2.2
contourpy==1.3.1
cryptography==46.0.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
executing==2.1.0
fastapi==0.116.2
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.55.3
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.8.0
fsspec==2025.7.0
h11==0.14.0
h2==4.3.0
hf-xet==1.1.5
hpack==4.1.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.34.3
hyperframe==6.1.0
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
langchain-classic==1.0.0
langchain-community==0.4
langchain-core==1.0.0
langchain-text-splitters==1.0.0
langsmith==0.4.37
lxml==5.3.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mistune==3.1.0
mpmath==1.3.0
multidict==6.7.0
multitasking==0.0.11
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.1.3
orjson==3.11.3
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
peewee==3.17.8
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
postgrest==2.19.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.4.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.2.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
rake-nltk==1.0.6
realtime==2.19.0
referencing==0.35.1
regex==2025.7.29
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
safetensors==0.5.3
scikit-learn==1.5.2
scipy==1.15.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.47.2
storage3==2.19.0
StrEnum==0.4.15
stripe==13.2.0
supabase==2.19.0
supabase-auth==2.19.0
supabase-functions==2.19.0
sympy==1.13.3
tenacity==9.1.2
terminado==0.18.1
textblob==0.19.0
tensorflow
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.21.4
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.1
types-python-dateutil==2.9.0.20241206
typing==3.7.4.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.2
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.35.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.13
yarl==1.22.0
yfinance==0.2.51
zstandard==0.25.0
PyPDF2
langchain_chroma
pypdf
unstructured
pdfminer
langchain-ollama
pdfminer.six 
pypdfium2 
langchain-community
pillow-heif
